codeguardian_ai/requirements.txt
# Core Flask dependencies
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3

# Database drivers
psycopg2-binary
redis

# Authentication and security
PyJWT
bcrypt
cryptography

# HTTP client and API integrations
requests
httpx
aiohttp

# Data processing and validation
pydantic
marshmallow
jsonschema

# Async support
asyncpg
aioredis

# Production server
gunicorn
gevent

# Monitoring and observability
prometheus-client
structlog
python-json-logger

# Development and testing
pytest
pytest-asyncio
pytest-cov
black
flake8
mypy

# Utilities
python-dotenv==1.0.0
click==8.2.1
blinker==1.9.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
greenlet==3.2.3

# System monitoring
psutil

# Code analysis and security
bandit

# Performance optimization
cachetools

# Configuration management
pyyaml
toml

